MWPW-159028: MAS doc for merch-card

+++ libs/features/mas/web-components/test/merch-card.mini-compare.test.html.js
@@ -52,7 +52,7 @@ runTests(async () => {
                         const el =
                             miniCompareChart.shadowRoot.querySelector(selector);
                         if (!el) return 0;
-                        return parseFloat(window.getComputedStyle(el).height);
+                        return parseInt(window.getComputedStyle(el).height);
                     })
                     .join(',');
                 return heights;
